Remove redundant job status queries.
If I'm understanding things correctly, all we need to do in jstaa is to
query for external_id's with unpropagated job status updates and post
them to the apps endpoint. There was a bunch of extra logic in there
from when jstaa had to do more, so it's been removed.